Corrected accessor for JSON data and made global constant.
Not all the line totals were summing correctly in the console so
I corrected the way the data structure was traversed (and modified
the data structure itself slightly at the top).

I moved the function outside of the render dispatch callback so
as to ensure it is self-contained/modular and potentially ready for
unit-testing later.