Added basic processing and rendering for simple file list.
Having created a tree data structure for the folder/file hierarchy
represented by the original JSON files, I have opted to try
rendering just the files as they are listed in the JSON data
to get a basic iternation of the visualisation up.

The current rendering can display both data sources but there is
no transition away from one to the other -- both display at once,
once the second one is selected.

Also the larger dataset is not readable at present. The height of
the chart needs more consideration.