Enabled basic directory exploration.
The click event allows for navigating through directories.

However, it is still not obvious to the user which directory they
are in, or how it is one gets back to other areas. So this work
is mainly a proof that it works but some of the behaviour will need
to change.