Moved data files into main directory for GistHub/Blocks demo.
Ideally the datafiles would be stored neatly in their own
subfolder. Unfortunately, GistHub does not allow creation of
subfolders within its repositories and seeing as I want this
repo to be available on GistHub as well as GitHub so that I can
show the work on my Bl.ocks portfolio, I've had to move the files
up into the main repository.